add ScriptEngine loadModuleSource add ScriptEngine loadModuleBinary update pybind

+++ src/System/SDLSoundSystem/SDLSoundSystem.cpp
@@ -73,7 +73,7 @@ namespace Menge
 
         m_countChannels = channels;
 
-        for( uint32_t i = 0; i != m_countChannels; ++i )
+        for( int i = 0; i != m_countChannels; ++i )
         {
             m_freeChannels[i] = true;
         }
@@ -97,7 +97,7 @@ namespace Menge
     //////////////////////////////////////////////////////////////////////////
     void SDLSoundSystem::update()
     {
-        for( uint32_t i = 0; i != m_countChannels; ++i )
+        for( int i = 0; i != m_countChannels; ++i )
         {
             bool free = m_freeChannels[i];
             
@@ -179,9 +179,9 @@ namespace Menge
         return soundSource;
     }
     //////////////////////////////////////////////////////////////////////////
-	uint32_t SDLSoundSystem::findFreeChannel()
+	int SDLSoundSystem::findFreeChannel()
     {
-        for( uint32_t i = 0; i < m_countChannels; ++i )
+        for( int i = 0; i < m_countChannels; ++i )
         {
             bool free = m_freeChannels[i];
 
@@ -195,10 +195,10 @@ namespace Menge
             return i;
         }
 
-        return (uint32_t )-1;
+        return -1;
     }
     //////////////////////////////////////////////////////////////////////////
-    void SDLSoundSystem::freeChannel( uint32_t channel)
+    void SDLSoundSystem::freeChannel( int channel)
     {
         if( channel >= 0 && channel < m_countChannels )
         {
